requirements.txt

alembic==1.0.11
amqp==2.4.1
aniso8601==4.1.0
asn1crypto==0.24.0
Authlib==0.14.1
billiard==3.5.0.5
blinker==1.4
celery==4.2.1
certifi==2019.6.16
cffi==1.12.0
chardet==3.0.4
Click==7.0
cryptography==2.5
Django==2.2.5
dominate==2.3.5
flasgger==0.9.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-Cors==3.0.8
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
ForgeryPy==0.1
gevent==1.4.0
greenlet==0.4.15
grpcio==1.24.1
grpcio-tools==1.24.1
gunicorn==19.9.0
httpie==0.9.9
idna==2.8
image==1.5.27
influxdb==5.2.3
itsdangerous==0.24
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==2.6.0
kombu==4.3.0
Mako==1.1.0
MarkupSafe==1.1.1
mistune==0.8.4
Pillow==6.0.0
protobuf==3.10.0
pycparser==2.19
pycryptodome==3.7.3
Pygments==2.2.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
PyYAML==5.1.2
qrcode==6.1
redis==3.3.6
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.6
sqlparse==0.3.0
urllib3==1.25.3
vine==1.2.0
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
xlrd==1.1.0
xlwt==1.3.0
